+++ src/renderer/VertexBuffer.h
@@ -1,10 +1,12 @@
 #pragma once
 
+#include <cstddef>
+
 class VertexBuffer {
 private:
     unsigned int m_rendererID;
 public:
-    VertexBuffer(const void* data, unsigned int size);
+    VertexBuffer(const void* data, size_t size);
     ~VertexBuffer();
 
     void bind() const;
